refactor(test): P5-1 extract Canvas2D bridge-fns + Yoga borderWidth
Two coherent extractions from test_widget_bridge.cpp:

1. test_widget_bridge_canvas2d_bridge_fns.cpp (10 TEST_CASEs, 66
   assertions, 446 lines) — bundles three closely related Canvas2D
   bridge entry-points:
     * pulp #1434 canvasSetFontFull bridge fn
     * pulp #1522 fillRule arg threading through canvasFill/canvasClip
     * pulp #1520 canvasSetDirection / canvasSetFilter bridge fns

2. test_widget_bridge_yoga_border.cpp (4 TEST_CASEs, 26 assertions,
   189 lines) — pulp #1543 Yoga borderWidth wiring
   (YGNodeStyleSetBorder integration with Yoga 3.x default
   box-sizing).

Also removes the now-unused canvasFromBridge helper from the parent
TU (all callers extracted).

test_widget_bridge.cpp: 4,444 → 3,862 (-582 lines, under 4k)
